refactor: reorganize goal domain into Data/UI packages
