+++ app/src/main/java/com/example/daybreak/Data/GoalItemData.kt
@@ -1,4 +1,4 @@
-package com.example.daybreak
+package com.example.daybreak.Data
 
 data class GoalItemData(
     val content: String // 목표 내용 텍스트
