--- app/src/main/java/com/example/daybreak/GoalDetailAdapter.kt
@@ -1,3 +1,0 @@
-package com.example.daybreak.adapter // 패키지명은 프로젝트에 맞게 확인하세요
-
-
