+++ app/src/main/java/com/example/daybreak/UI/dialog/ToastDialogFragment.kt
@@ -1,6 +1,5 @@
-package com.example.daybreak.UI.dialog
+package com.example.daybreak
 
-import android.R
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
@@ -45,13 +44,24 @@ class ToastDialogFragment(private val message: String, private val iconResId: In
             setGravity(Gravity.TOP)
             //위 마진 조절
             val params = attributes
-            params.y = 100
-            params.windowAnimations = R.style.Animation_Toast
+            params.y = 40
+            params.windowAnimations = android.R.style.Animation_Toast
             attributes = params
         }
 
         Handler(Looper.getMainLooper()).postDelayed({
             if(isAdded) dismiss()
         }, 3000)
+    }
+    override fun onResume() {
+        super.onResume()
+        dialog?.window?.let { window ->
+            val params = window.attributes
+            // 화면 너비의 90%로 설정
+            params.width = (resources.displayMetrics.widthPixels * 0.9).toInt()
+            // 높이는 내용물에 맞게 자동 조절
+            params.height = android.view.ViewGroup.LayoutParams.WRAP_CONTENT
+            window.attributes = params
+        }
     }
 }
