+++ app/src/main/java/com/example/daybreak/UI/main/MainActivity.kt
@@ -7,8 +7,8 @@ import androidx.activity.addCallback
 import androidx.appcompat.app.AppCompatActivity
 import com.example.daybreak.GoalFlowActivity
 import com.example.daybreak.R
+import com.example.daybreak.ToastDialogFragment
 import com.example.daybreak.UI.adapter.MainVPAdapter
-import com.example.daybreak.UI.dialog.ToastDialogFragment
 import com.example.daybreak.UI.fragment.SettingFragment
 import com.example.daybreak.databinding.ActivityMainBinding
 import com.google.android.material.tabs.TabLayoutMediator
